app/requirements.txt
asgiref==3.4.1
astroid==2.8.6
async-generator==1.10
attrs==21.2.0
black==21.11b1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.50.0
django-debug-toolbar==3.2.2
django-dumpdata-one==0.8.5
django-heroku==0.3.1
flake8==4.0.1
gunicorn==20.1.0
h11==0.12.0
idna==3.3
importlib-metadata==4.2.0
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.0
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyOpenSSL==21.0.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
trio==0.19.0
trio-websocket==0.9.2
typed-ast==1.5.0
typing_extensions==4.0.0
urllib3==1.26.7
whitenoise==6.0.0
wrapt==1.13.3
wsproto==1.0.0
zipp==3.6.0
